# services/music-stream/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Private Music Playback API",
    "version": "1.0.0",
    "description": "One current contract for temporary music playback authorization."
  },
  "servers": [
    {
      "url": "https://audio.tyemirov.net"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
    "/api/playback-grants": {
      "post": {
        "operationId": "createPlaybackGrant",
        "security": [
          {},
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "An exact permitted website origin."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrant"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A playback grant was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Grant"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "The grant resource path.",
                "schema": {
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "The Secure, HttpOnly, SameSite=Strict session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Failure"
          },
          "403": {
            "$ref": "#/components/responses/Failure"
          },
          "404": {
            "$ref": "#/components/responses/Failure"
          },
          "405": {
            "$ref": "#/components/responses/Failure"
          },
          "409": {
            "$ref": "#/components/responses/Failure"
          },
          "413": {
            "$ref": "#/components/responses/Failure"
          },
          "415": {
            "$ref": "#/components/responses/Failure"
          },
          "429": {
            "$ref": "#/components/responses/Failure"
          },
          "503": {
            "$ref": "#/components/responses/Failure"
          }
        }
      }
    },
    "/api/playback-grants/{grantId}": {
      "parameters": [
        {
          "name": "grantId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{22}$"
          }
        }
      ],
      "get": {
        "operationId": "readPlaybackGrant",
        "responses": {
          "200": {
            "description": "The current playback grant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Grant"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "description": "The response has no-store cache policy.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Failure"
          },
          "401": {
            "$ref": "#/components/responses/Failure"
          },
          "403": {
            "$ref": "#/components/responses/Failure"
          },
          "404": {
            "$ref": "#/components/responses/Failure"
          },
          "405": {
            "$ref": "#/components/responses/Failure"
          },
          "410": {
            "$ref": "#/components/responses/Failure"
          },
          "413": {
            "$ref": "#/components/responses/Failure"
          },
          "415": {
            "$ref": "#/components/responses/Failure"
          },
          "429": {
            "$ref": "#/components/responses/Failure"
          },
          "503": {
            "$ref": "#/components/responses/Failure"
          }
        }
      },
      "delete": {
        "operationId": "deletePlaybackGrant",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "An exact permitted website origin."
          }
        ],
        "responses": {
          "204": {
            "description": "The grant is revoked."
          },
          "400": {
            "$ref": "#/components/responses/Failure"
          },
          "401": {
            "$ref": "#/components/responses/Failure"
          },
          "403": {
            "$ref": "#/components/responses/Failure"
          },
          "404": {
            "$ref": "#/components/responses/Failure"
          },
          "405": {
            "$ref": "#/components/responses/Failure"
          },
          "410": {
            "$ref": "#/components/responses/Failure"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-music-session",
        "description": "The host-only session cookie belongs to the media origin."
      }
    },
    "responses": {
      "Failure": {
        "description": "A typed request failure.",
        "headers": {
          "Cache-Control": {
            "description": "The response has no-store cache policy.",
            "schema": {
              "type": "string"
            }
          },
          "Retry-After": {
            "description": "For 429, wait this number of seconds.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          "X-Request-ID": {
            "description": "The request identifier.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "CreateGrant": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "trackId"
        ],
        "properties": {
          "trackId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,79}$"
          }
        }
      },
...
      "Grant": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "grantId",
          "trackId",
          "playlistUrl",
          "durationMs",
          "serverTime",
          "expiresAt"
        ],
        "properties": {
          "grantId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{22}$"
          },
          "trackId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,79}$"
          },
          "playlistUrl": {
            "type": "string",
            "format": "uri",
            "pattern": "^https://[^/]+/hls/[A-Za-z0-9_-]{22}/[a-f0-9]{64}/index\\.m3u8$"
          },
          "durationMs": {
            "type": "integer",
            "minimum": 1000,
            "maximum": 7200250
          },
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "session_required",
                  "origin_denied",
                  "not_found",
                  "grant_expired",
                  "track_unavailable",
                  "grant_limit",
                  "rate_limited",
                  "media_unavailable",
                  "invalid_request",
                  "json_required",
                  "body_too_large",
                  "method_not_allowed",
                  "headers_denied"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}